Separated blob class
